Debugger no longer experimental?
I didn't notice "experimental" in the docs, only in the README, and just thought it might be outdated with version 1.0 out (and seeing debug issue closed). If not ignore or close. All software may have bugs or missing features, but seeing "experimental" may put people off.

In general I would like to know where VSCode extension has feature parity with Juno (and well anything else used for Julia) or is better.

Such as for the debugger, linter, plotting etc. All considered non-experimental? I kind of liked the way in Juno to highlight code and run and getting checkboxes. Is it on the agenda or already there somewhere with non-default option?

I do not really care what implements the debugger if it just works. Is there one built-in go to option, this one (or as an option?):

https://github.com/judy-vscode/Adapter

https://discourse.julialang.org/t/tips-for-debugging-in-julia-vs-code-while-using-large-packages/43371/7