Enhance the custom sysimage building
Exclude the development packages inside the profile as they are
often subject to changes, hence should not be made static.
Adds the ability to set the precompile_statements_file, and the
precompile_execution_file arguments to the `create_sysimage` function.
These settings will be read from a `vscode-compileenv.tmol` file's
`sysimage` table.
This file should be inside  the package directory.
Additional excluded files can be added under the `excluded_packages`
field as well.